requirements-predicting.txt
blinker==1.8.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.53.1
gunicorn==23.0.0
importlib_metadata==8.4.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==1.26.4
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
tflite-runtime==2.14.0
Werkzeug==3.0.4
zipp==3.20.1
